requirements.txt
torch~=2.5.1+cu118
pillow~=10.3.0
diffusers~=0.31.0
transformers~=4.46.2